# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: MetanoiaDocs API
  version: '0.1'
  description: |
    Everything here is same-origin with the app. Two ways to authenticate:

    - **A session cookie** (`md_session`), which is what the browser uses.
    - **A personal access token** as `Authorization: Bearer <token>`, which is
      what everything else uses. Make one in Settings → API tokens. A token
      carries exactly the access of the person who made it — there are no
      scopes, because there is nothing a token should be able to do that its
      owner cannot.

    Times are ISO-8601. Dates on tasks are `YYYY-MM-DD` — a due date is a day,
    not an instant. Errors are `{ "error": "<sentence>" }` with a 4xx or 5xx
    status; the sentence is written to be shown to a person.
  license:
    name: MIT
    url: https://github.com/Xephyr-Labs/metanoiadocs/blob/main/LICENSE
servers:
  - url: http://localhost:8092
    description: A local instance from the quick start
  - url: https://docs.example.com
    description: Your instance
security:
  - bearerAuth: []
...
paths:
...
  /api/projects/{id}/automations:
    get:
      tags: [Automations]
      summary: A database's rules
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Rules, in the order they run
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Automation' } }
    post:
      tags: [Automations]
      summary: Create a rule
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                trigger: { type: string, enum: [status, manual], default: status }
                value:
                  type: [string, 'null']
                  description: The status a task has to enter, for a `status` rule.
                actions:
                  type: array
                  items: { $ref: '#/components/schemas/Action' }
      responses:
        '200':
          description: The new rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Automation' }
  /api/automations/{id}:
    patch:
      tags: [Automations]
      summary: Change a rule
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                trigger: { type: string, enum: [status, manual] }
                value: { type: [string, 'null'] }
                actions: { type: array, items: { $ref: '#/components/schemas/Action' } }
                active: { type: boolean }
                position: { type: integer }
      responses:
        '200':
          description: The rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Automation' }
    delete:
      tags: [Automations]
      summary: Delete a rule
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /api/tasks/{id}/automations/{automationId}/run:
    post:
      tags: [Automations]
      summary: Run a rule on one task, now
      description: What makes a rule a quick action. Works for either trigger kind.
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: automationId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: What actually changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { const: true }
                  applied: { type: array, items: { type: object } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal access token from Settings → API tokens.
    cookieAuth:
      type: apiKey
      in: cookie
      name: md_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Ok:
      description: Done
      content:
        application/json:
          schema: { type: object, properties: { ok: { const: true } } }
...
  schemas:
...
    Action:
      type: object
      description: One thing a rule does.
      required: [type]
      properties:
        type:
          type: string
          enum: [assign, status, kind, sprint, priority, points, progress]
        userIds:
          type: array
          items: { type: string }
          description: For `assign` only.
        value:
          description: |
            For everything but `assign`. A sprint action takes a sprint id,
            `"active"` for whichever sprint is running, or null for the backlog.
    Automation:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        trigger_kind: { type: string, enum: [status, manual] }
        trigger_value: { type: [string, 'null'] }
        actions: { type: array, items: { $ref: '#/components/schemas/Action' } }
        active: { type: boolean }
        position: { type: integer }
...